pyproject.toml
[tool.poetry]
name = "simbatolls"
version = "1"
description = ""
authors = ["Bardia"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.1"
flask = "^3.0.3"
plotly = "^5.20.0"
numpy = "^1.26.4"
datetime = "^5.5"
werkzeug = "^3.0.2"
openpyxl = "^3.1.2"
ipykernel = "^6.29.4"
pandasql = "^0.7.3"
matplotlib = "^3.8.4"
pyqt6 = "^6.6.1"
xlrd = "^2.0.1"
temp = "^2020.7.2"
gunicorn = "^21.2.0"
flask-login = "^0.6.3"
flask-sqlalchemy = "^3.1.1"
flask-bcrypt = "^1.0.1"
app = "^0.0.1"
bcrypt = "^4.1.2"
psycopg2-binary = "^2.9.9"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
